re-organize assets and add miguel teheran as sponsor
