+++ src/assets/pictures/qr-code-nequi.png
