+++ src/components/donate/donate.scss
@@ -45,7 +45,7 @@
         font-weight: 700;
 
         &-decor {
-          background-image: url("../../assets/feather.svg");
+          background-image: url("../../assets/illustrations/feather.svg");
           background-position: center;
           background-size: contain;
           background-repeat: no-repeat;
