+++ src/components/donateDetails/donateDetails.scss
@@ -72,7 +72,7 @@
     &::after {
       content: "";
       position: absolute;
-      background-image: url("../../assets/feather-pink.svg");
+      background-image: url("../../assets/illustrations/feather-pink.svg");
       background-position: center;
       background-size: contain;
       background-repeat: no-repeat;
