+++ src/components/event/event.tsx
@@ -1,7 +1,7 @@
 import { Link } from "react-router-dom";
 import { useTranslation } from "react-i18next";
 import { Button } from "@/components/ui/button";
-import logoIconOnly from "@/assets/logo-icon-only.svg";
+import logoIconOnly from "@/assets/guardianes-logos/logo-icon-only.svg";
 import "./event.scss";
 
 export const Event = (): JSX.Element => {
