+++ src/components/story/story.scss
@@ -23,7 +23,7 @@
     &-image {
       @apply max-h-72 md:max-h-96;
 
-      mask: url("../../assets/mask-blob.svg");
+      mask: url("../../assets/illustrations/mask-blob.svg");
     }
   }
 }
